21.12.09

+++ src/main/java/WebGallery/Gallery/service/AuthorService.java
@@ -46,15 +46,17 @@ public class AuthorService {
         try {
             Author author = authorRepository.findByGno(gno);
             if(author == null) return response.fail("작가가 존재하지 않습니다.", HttpStatus.BAD_REQUEST);
+            Guest guest = guestRepository.findByGno(gno);
+            if(guest == null) return response.fail("게스트가 존재하지 않습니다.", HttpStatus.BAD_REQUEST);
 
             authorRepository.delete(author);
             log.info("author delete success");
-            Guest guest = guestRepository.findByGno(gno);
             guest.changeRole(Role.GUEST);
             guestRepository.save(guest);
             log.info("Role Change");
             map.put("delete", "success");
-            return response.success(map,"작가 삭제 성공", HttpStatus.OK);
+            return response.success(map, "작가 삭제 성공", HttpStatus.OK);
+
         } catch (IllegalArgumentException exception){
             exception.printStackTrace();
         }
